requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.19.0
requests>=2.31.0
python-dotenv>=1.0.0
fredapi>=0.5.1
schedule>=1.2.1
python-telegram-bot>=20.7
pytest>=8.0.0
